+++ src-docs/src/components/with_theme/theme_context.tsx
@@ -6,7 +6,7 @@ import { applyTheme } from '../../services';
 const THEME_NAMES = EUI_THEMES.map(({ value }) => value);
 
 const defaultState = {
-  theme: THEME_NAMES[0],
+  theme: THEME_NAMES[2],
   changeTheme: (themeValue: EUI_THEME['value']) => {
     applyTheme(themeValue);
   },